login, signup page, started CRUD operations
