+++ client/package.json
@@ -22,7 +22,8 @@
     "@vueuse/core": "^13.6.0",
     "primeicons": "^7.0.0",
     "primevue": "^4.3.7",
-    "vue": "^3.5.18"
+    "vue": "^3.5.18",
+    "vue-router": "^4.5.1"
   },
   "devDependencies": {
     "@tsconfig/node22": "^22.0.2",
