+++ client/src/main.ts
@@ -5,6 +5,7 @@ import App from './App.vue';
 import PrimeVue from 'primevue/config';
 import { DatePicker } from 'primevue';
 import Aura from '@primeuix/themes/aura';
+import router from './router';
 
 const app = createApp(App);
 
@@ -14,4 +15,4 @@ app.use(PrimeVue, {
     }
 });
 app.component('DatePicker',DatePicker);
-app.mount('#app');
+app.use(router).mount('#app');
