+++ client/src/views/LoginView.vue
@@ -1,13 +1,7 @@
 <script setup lang="ts">
-import { ref, defineProps } from 'vue';
-import BaseButton from './BaseButton.vue';
+import { ref } from 'vue';
+import BaseButton from '../components/BaseButton.vue';
 
-const props = defineProps<{
-  mode: string;
-}>();
-
-const firstName = ref<string>('');
-const lastName = ref<string>('');
 const email = ref<string>('');
 const password = ref<string>('');
 </script>
@@ -15,16 +9,6 @@ const password = ref<string>('');
 <template>
   <form>
     <div class="bg-[#efefef] px-6 py-4 font-sans border-0 rounded-2xl text-lg">
-      <div v-show="mode === 'signin'">
-        <div class="flex items-center">
-          <label>First name: </label>
-          <input v-model="firstName" /><br />
-        </div>
-        <div class="flex items-center">
-          <label>Last name: </label>
-          <input v-model="lastName" /><br />
-        </div>
-      </div>
       <div class="flex items-center">
           <label>Email address: </label>
           <input v-model="email" /><br />
@@ -35,8 +19,7 @@ const password = ref<string>('');
         </div>
       <br />
       <div class="flex justify-end">
-        <BaseButton v-if="mode === 'signin'" html-type="submit">Sign in</BaseButton>
-        <BaseButton v-else html-type="submit">Login</BaseButton>
+        <BaseButton html-type="submit">Login</BaseButton>
       </div>
     </div>
   </form>
