+++ server/src/routes/tasks.ts
@@ -1,15 +1,13 @@
-import { Router, Request, Response} from 'express'
-import { Task } from '../models/Task';
-
+import { Router } from 'express'
+import { getTasks, createTask, updateTaskImportance, updateTaskState } from '../controllers/taskController'
 const router = Router();
 
-router.get("/", async(req: Request, res: Response) => {
-  try {
-    const tasks = await Task.find();
-    res.json(tasks);
-  } catch(err) {
-    res.sendStatus(500);
-  }
-});
+router.get("/", getTasks);
+
+router.post("/", createTask);
+
+router.put("/:_id/importance", updateTaskImportance);
+
+router.put("/:_id/completed", updateTaskState);
 
 export default router;
